+++ src/debvader/deblend/field_deblender.py
@@ -1,7 +1,12 @@
+import numpy as np
+import scipy
+import pandas as pd
+
 from debvader.detect.detection import detect_objects
 from debvader.extract.extraction import extract_cutouts
 from debvader.deblend_cutout.metrics import mse
 from debvader.deblend_cutout.optimization import position_optimization
+from debvader.deblend_cutout.deblender import deblend
 
 
 class DeblendField:
@@ -12,7 +17,7 @@ class DeblendField:
         cutout_size=59,
         nb_of_bands=6,
         epistemic_uncertainty_estimation=False,
-        normalised=False,
+        normalise=False,
     ):
         """
         to initialize
@@ -23,7 +28,7 @@ class DeblendField:
             cutout_size: size of the stamps
             nb_of_bands: number of filters in the image
             epistemic_uncertainty_estimation: boolean to indication if expestemic uncertainity extimation is to be done.
-            normalised: boolean to indicate if images need to be normalised
+            normalise: boolean to indicate if images need to be normalise
         """
 
         self.net = net
@@ -32,7 +37,7 @@ class DeblendField:
         self.cutout_size = cutout_size
         self.nb_of_bands = nb_of_bands
         self.epistemic_uncertainty_estimation = epistemic_uncertainty_estimation
-        self.normalised = normalised
+        self.normalise = normalise
         self.nb_of_detected_objects = []
         self.nb_of_deblended_galaxies = []
         self.res_deblend = None
@@ -253,7 +258,7 @@ class DeblendField:
         # Deblend the cutouts around the detected galaxies. If needed, create the cutouts.
         if isinstance(cutout_images, np.ndarray):
             output_images_mean, output_images_distribution = deblend(
-                self.net, cutout_images, normalised=self.normalised
+                self.net, cutout_images, normalise=self.normalise
             )
             list_idx = list(range(0, len(output_images_mean)))
         else:
@@ -265,7 +270,7 @@ class DeblendField:
                 self.nb_of_bands,
             )
             output_images_mean, output_images_distribution = deblend(
-                self.net, cutout_images[list_idx], normalised=self.normalised
+                self.net, cutout_images[list_idx], normalise=self.normalise
             )
         if list_idx == []:
             print("No galaxy deblended. End of the iterative procedure.")
@@ -301,7 +306,7 @@ class DeblendField:
                         deblend(
                             self.net,
                             np.array([cutout_images[k]] * 100),
-                            normalised=self.normalised,
+                            normalise=self.normalise,
                         )[0],
                         axis=0,
                     )
@@ -376,135 +381,3 @@ class DeblendField:
 
         return self.res_deblend
 
-    def iterative_deblending(
-        self,
-        galaxy_distances_to_center=None,
-        cutout_images=None,
-        optimise_positions=False,
-        epistemic_criterion=100.0,
-        mse_criterion=100.0,
-    ):
-        """
-        Do the iterative deblending of a scene
-        paramters:
-            galaxy_distances_to_center: distances of the galaxies to deblend from the center of the field. In pixels.
-            cutout_images: stamps centered on the galaxies to deblend
-            optimise_position: boolean to indicate if the user wants to use the scipy optimize package to optimise the position of the galaxy
-            epistemic_criterion: cut for epistemic uncertainity to get rid of bad predictions
-            mse_criterion: cut for mse_criterion to get rid of bad predictions
-            normalised: boolean to indicate if images need to be normalised
-        """
-
-        # do the first step of deblending
-        field_image = self.field_image.copy()
-        res_step = self.deblending_step(
-            field_image,
-            cutout_images=cutout_images,
-            optimise_positions=optimise_positions,
-            epistemic_criterion=epistemic_criterion,
-            mse_criterion=mse_criterion,
-        )
-        res_deblend = res_step
-
-        new_residual_field = self.get_residual_field()
-        self.mse += [mse(self.field_image, new_residual_field)]
-        shifts_previous = []
-        k = 1
-        diff_mse = -1
-
-        # Now iterate over
-        while len(res_step["shifts"]) > len(shifts_previous):
-
-            print(f"iteration {k}")
-            shifts_previous = res_step["shifts"]
-
-            prev_residual_field = new_residual_field
-
-            # deblending step will run detection and deblending on the residual field
-            res_step = self.deblending_step(
-                prev_residual_field,
-                cutout_images=None,
-                optimise_positions=optimise_positions,
-                mse_criterion=mse_criterion,
-            )
-
-            # compute the MSE after this iteration step
-            new_residual_field = self.get_residual_field()
-            self.mse += [mse(prev_residual_field, new_residual_field)]
-            # field_img_save, field_image, denoised_field, denoised_field_std, denoised_field_epistemic, cutout_images, output_images_mean, output_images_distribution, shifts, galaxy_distances_to_center, mse_step = deblending_step(net, field_img_init, detection_up_to_k, cutout_images = None, cutout_size = cutout_size, nb_of_bands = nb_of_bands, optimise_positions=optimise_positions, epistemic_uncertainty_estimation=epistemic_uncertainty_estimation, epistemic_criterion=epistemic_criterion, mse_criterion=mse_criterion, normalised=normalised)
-            # field_img_init=field_img_save.copy()
-
-            if res_step["list_idx"] is None:
-                break
-
-            res_deblend = np.concatenate([res_deblend, res_step])
-            k += 1
-
-            print(
-                f"{sum(self.nb_of_deblended_galaxies)} galaxies found up to this step."
-            )
-            print(
-                f"deta_mse = {diff_mse}, mse_iteration = "
-                + str(self.mse[-1])
-                + " and mse_previous_step = "
-                + str(self.mse[-2])
-            )
-
-        print("converged !")
-
-        self.res_deblend = res_deblend
-
-        return self.res_deblend
-
-    def deblending_step(
-        self,
-        field_image,
-        cutout_images=None,
-        optimise_positions=False,
-        epistemic_criterion=100.0,
-        mse_criterion=100.0,
-    ):
-        """
-        One step of the iterative procedure called within iterative_procedure.
-
-        paramters:
-            field_image: image of the field to deblend
-            cutout_images: stamps centered on the galaxies to deblend
-            optimise_position: boolean to indicate if the user wants to use the scipy optimize package to optimise the position of the galaxy
-            epistemic_criterion: cut for epistemic uncertainity to get rid of bad predictions
-            mse_criterion: cut for mse_criterion to get rid of bad predictions
-        """
-        detection_k = detect_objects(field_image)
-        # Avoid to have several detection at the same location
-
-        # TODO: Fix this part to get rid of false detections.
-        # Ideally call a remove_residual_detection function which can be developed later!
-
-        # if isinstance(galaxy_distances_to_center_total, np.ndarray):
-        #    for i in range (len(detection_k)):
-        #        if detection_k[i] in galaxy_distances_to_center_total: TODO: Does this make sense??
-        #            idx_to_remove.append(i)
-        #    detection_k = np.delete(detection_k, idx_to_remove, axis = 0)
-
-        res_step = self.deblend_field(
-            field_image=field_image,
-            galaxy_distances_to_center=detection_k,
-            cutout_images=cutout_images,
-            optimise_positions=optimise_positions,
-            epistemic_criterion=epistemic_criterion,
-            mse_criterion=mse_criterion,
-        )
-
-        # field_img_save, field_image, denoised_field, denoised_field_std, denoised_field_epistemic, cutout_images, output_images_mean, output_images_distribution, shifts, list_idx, nb_of_galaxies_in_deblended_field = deblend_field(net, field_image, detection_k, cutout_images = cutout_images, cutout_size = cutout_size, nb_of_bands = nb_of_bands, optimise_positions=optimise_positions, epistemic_uncertainty_estimation=epistemic_uncertainty_estimation, epistemic_criterion=epistemic_criterion, mse_criterion=mse_criterion, normalised=normalised)
-        if len(res_step["list_idx"]) == 0:
-            print("No more galaxies found")
-            return self.res_deblend
-
-        res_step["list_idx"] += (
-            sum(self.nb_of_deblended_galaxies) - self.nb_of_deblended_galaxies[-1]
-        )
-
-        print(f"Deblend {self.nb_of_deblended_galaxies[-1]} more galaxy(ies)")
-        # detection_confirmed = np.zeros((len(res_step["list_idx"]), 2))
-
-        return res_step
