+++ src/debvader/deblend_cutout/deblender.py
@@ -1,25 +1,25 @@
-import numpy as np
-import pandas as pd
-import scipy
 import tensorflow as tf
 
+from debvader.normalize.normalize import normalize_non_linear, denormalize_non_linear
 
 
-def deblend(net, images, normalised=False):
+def deblend(net, images, normalise=False):
     """
     Deblend the image using the network
     parameters:
         net: neural network used to do the deblending
         images: array of images. It can contain only one image
-        normalised: boolean to indicate if images need to be normalised
+        normalise: boolean to indicate if images need to be normalised
     """
-    if normalised:
+    if normalise:
         # Normalize input images
-        images_normed = np.tanh(np.arcsinh(images))
-        # Denorm output images
-        images = np.sinh(np.arctanh(images_normed))
+        images = normalize_non_linear(images)
 
-    return net(tf.cast(images, tf.float32)).mean().numpy(), net(
-        tf.cast(images, tf.float32)
-    )
+    outimg = net(tf.cast(images, tf.float32))
+
+    if normalise:
+        # Denorm output images
+        outimg = denormalize_non_linear(outimg)
+
+    return (outimg.mean().numpy(), outimg)
